backend/pyproject.toml
[tool.poetry]
name = "docwiz-backend"
version = "0.1.0"
description = "DocWiz surgical visualization and cost estimation backend"
authors = ["DocWiz Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
# Firebase/Firestore (Google Cloud Database)
firebase-admin = "^6.2.0"
google-cloud-firestore = "^2.13.0"
# Redis and Celery
redis = "^5.0.0"
celery = "^5.3.0"
# WebSocket support
websockets = ">=13.0"
# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"
cryptography = "^41.0.0"
email-validator = "^2.3.0"
# Google Cloud Services
google-cloud-storage = "^2.10.0"
google-genai = "^1.0.0"
# Vector Database
qdrant-client = "^1.6.0"
# HTTP and Image Processing
httpx = "^0.25.0"
pillow = "^10.1.0"
reportlab = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
hypothesis = "^6.92.0"
black = "^23.10.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
isort = "^5.12.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "property_test: Property-based tests using Hypothesis",
    "integration: Integration tests requiring external services",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
